refactor!: change Message class to CommandFeedback class
